requirements.txt
torch>=1.7.0
torchvision==0.15.2
Pillow>=8.0.0
numpy>=1.19.2
scikit-image==0.20.0
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
gunicorn>=20.1.0 